moving fit_langau method from Pad- to TelescopeAnalysis class
